handle disagg with no sources without error. needs toshiapi update for dummy hazardsolution ID

+++ runzi/execute/openquake/execute_openquake.py
@@ -44,51 +44,61 @@ def execute_openquake(configfile, task_no, toshi_task_id):
         #
         cmd = ['oq', 'engine', '--run', f'{configfile}', '-L',  f'{logfile}']
         log.info(f'cmd 1: {cmd}')
-        subprocess.check_call(cmd)
+        oq_out = subprocess.run(cmd, capture_output=True)
+        log.info(oq_out.stdout.decode('UTF-8'))
+        log.info(oq_out.stderr.decode('UTF-8'))
+        if 'Filtered away all ruptures??' not in oq_out.stderr.decode('UTF-8'):
 
-        def get_last_task():
-            """
-            root@tryharder-ubuntu:/app# oq engine --lhc
-            job_id |     status |          start_time |         description
-                 6 |   complete | 2022-03-29 01:12:16 | 35 sites, few periods
-            """
+            def get_last_task():
+                """
+                root@tryharder-ubuntu:/app# oq engine --lhc
+                job_id |     status |          start_time |         description
+                    6 |   complete | 2022-03-29 01:12:16 | 35 sites, few periods
+                """
 
-            cmd = ['oq', 'engine', '--lhc']
-            p = subprocess.Popen(cmd, stdout=subprocess.PIPE)
-            out, err = p.communicate()
+                cmd = ['oq', 'engine', '--lhc']
+                p = subprocess.Popen(cmd, stdout=subprocess.PIPE)
+                out, err = p.communicate()
 
-            fileish = io.StringIO()
-            fileish.write(out.decode())
-            fileish.seek(0)
+                fileish = io.StringIO()
+                fileish.write(out.decode())
+                fileish.seek(0)
 
-            fileish.readline() #consume header
-            #lines = fileish.readlines()
-            for line in fileish.readlines():
-                print(line)
-                task = int(line.split("|")[0])
+                fileish.readline() #consume header
+                #lines = fileish.readlines()
+                for line in fileish.readlines():
+                    print(line)
+                    task = int(line.split("|")[0])
 
-            return task
+                return task
 
-        #get the job ID
-        last_task = get_last_task()
-        oq_result['oq_calc_id'] = last_task
+            #get the job ID
+            last_task = get_last_task()
+            oq_result['oq_calc_id'] = last_task
 
 
-        #
-        #  oq engine --export-outputs 12 /WORKING/examples/output/PROD/34-sites-few-CRU+BG
-        #  cp /home/openquake/oqdata/calc_12.hdf5 /WORKING/examples/output/PROD
-        #
-        cmd = ['oq', 'engine', '--export-outputs', str(last_task), str(output_path)]
-        log.info(f'cmd 2: {cmd}')
-        subprocess.check_call(cmd, stdout=subprocess.DEVNULL)
-        oq_result['csv_archive'] = archive(output_path, Path(WORK_PATH, f'openquake_csv_archive-{toshi_task_id}.zip'))
+            #
+            #  oq engine --export-outputs 12 /WORKING/examples/output/PROD/34-sites-few-CRU+BG
+            #  cp /home/openquake/oqdata/calc_12.hdf5 /WORKING/examples/output/PROD
+            #
+            cmd = ['oq', 'engine', '--export-outputs', str(last_task), str(output_path)]
+            log.info(f'cmd 2: {cmd}')
+            subprocess.check_call(cmd, stdout=subprocess.DEVNULL)
+            oq_result['csv_archive'] = archive(output_path, Path(WORK_PATH, f'openquake_csv_archive-{toshi_task_id}.zip'))
 
-        #clean up export outputs
-        shutil.rmtree(output_path)
+            #clean up export outputs
+            shutil.rmtree(output_path)
 
-        OQDATA = "/home/openquake/oqdata"
-        hdf5_file = f"calc_{last_task}.hdf5"
-        oq_result['hdf5_archive'] = archive(Path(OQDATA, hdf5_file), Path(WORK_PATH, f'openquake_hdf5_archive-{toshi_task_id}.zip'))
+            OQDATA = "/home/openquake/oqdata"
+            hdf5_file = f"calc_{last_task}.hdf5"
+            oq_result['hdf5_archive'] = archive(Path(OQDATA, hdf5_file), Path(WORK_PATH, f'openquake_hdf5_archive-{toshi_task_id}.zip'))
+        else:
+            oq_result['csv_archive']=Path(WORK_PATH, f"spoof-{task_no}.csv_archive.zip")
+            oq_result['hdf5_archive']=Path(WORK_PATH, f"spoof-{task_no}.hdf5_archive.zip")
+            oq_result['csv_archive'].touch()
+            oq_result['hdf5_archive'].touch()
+            oq_result['no_result'] = True
+      
 
     except Exception as err:
         log.error(f"err: {err}")
